[MINOR] Removed FSUtils.makeBaseFileName without fileExt param
